refactor: move profile persistence helpers
