+++ connection_profile.go
@@ -10,6 +10,8 @@ import (
 
 	"github.com/BurntSushi/toml"
 	"github.com/zalando/go-keyring"
+
+	"github.com/marang/goemqutiti/internal/files"
 )
 
 // Profile defines a broker connection.
@@ -191,4 +193,58 @@ func LoadProfile(name, file string) (*Profile, error) {
 		return nil, fmt.Errorf("no connection profile available")
 	}
 	return p, nil
+}
+
+// saveConfig persists profiles and default selection to config.toml.
+func saveConfig(profiles []Profile, defaultName string) {
+	saved := loadState()
+	cfg := userConfig{
+		DefaultProfileName: defaultName,
+		Profiles:           profiles,
+		Saved:              make(map[string]persistedConn),
+	}
+	for k, v := range saved {
+		var topics []persistedTopic
+		for _, t := range v.Topics {
+			topics = append(topics, persistedTopic{Title: t.title, Active: t.subscribed})
+		}
+		var payloads []persistedPayload
+		for _, p := range v.Payloads {
+			payloads = append(payloads, persistedPayload{Topic: p.topic, Payload: p.payload})
+		}
+		cfg.Saved[k] = persistedConn{Topics: topics, Payloads: payloads}
+	}
+	writeConfig(cfg)
+}
+
+// savePasswordToKeyring stores a password in the system keyring.
+func savePasswordToKeyring(service, username, password string) {
+	if err := keyring.Set("emqutiti-"+service, username, password); err != nil {
+		fmt.Println("Error saving password to keyring:", err)
+	}
+}
+
+// deleteProfileData removes profile-specific persisted history and traces.
+func deleteProfileData(name string) {
+	os.RemoveAll(filepath.Join(files.DataDir(name), "history"))
+	os.RemoveAll(filepath.Join(files.DataDir(name), "traces"))
+}
+
+// persistProfileChange applies a profile update, saves config and keyring.
+func persistProfileChange(profiles *[]Profile, defaultName string, p Profile, idx int) {
+	plain := p.Password
+	if !p.FromEnv {
+		p.Password = "keyring:emqutiti-" + p.Name + "/" + p.Username
+	} else {
+		p.Password = ""
+	}
+	if idx >= 0 && idx < len(*profiles) {
+		(*profiles)[idx] = p
+	} else {
+		*profiles = append(*profiles, p)
+	}
+	saveConfig(*profiles, defaultName)
+	if !p.FromEnv {
+		savePasswordToKeyring(p.Name, p.Username, plain)
+	}
 }
